Refactor examples to use configureStore
