+++ examples/basic/src/index.js
@@ -1,25 +1,11 @@
 import { AppContainer } from 'react-hot-loader'
-import { applyMiddleware, compose, createStore } from 'redux'
-import { createBrowserHistory } from 'history'
-import { routerMiddleware } from 'connected-react-router'
 import { Provider } from 'react-redux'
 import React from 'react'
 import ReactDOM from 'react-dom'
 import App from './App'
-import rootReducer from './reducers'
+import configureStore, { history } from './configureStore'
 
-const history = createBrowserHistory()
-
-const composeEnhancer = window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__ || compose
-const store = createStore(
-  rootReducer(history),
-  composeEnhancer(
-    applyMiddleware(
-      routerMiddleware(history),
-    ),
-  ),
-)
-
+const store = configureStore()
 const render = () => {
   ReactDOM.render(
     <AppContainer>
@@ -38,10 +24,5 @@ if (module.hot) {
   // Reload components
   module.hot.accept('./App', () => {
     render()
-  })
-
-  // Reload reducers
-  module.hot.accept('./reducers', () => {
-    store.replaceReducer(rootReducer(history))
   })
 }
