+++ examples/immutable/src/index.js
@@ -1,28 +1,13 @@
 import { AppContainer } from 'react-hot-loader'
-import { applyMiddleware, compose, createStore } from 'redux'
-import { createBrowserHistory } from 'history'
-import { routerMiddleware } from 'connected-react-router/immutable'
 import { Provider } from 'react-redux'
 import Immutable from 'immutable'
 import React from 'react'
 import ReactDOM from 'react-dom'
 import App from './App'
-import rootReducer from './reducers'
-
-const history = createBrowserHistory()
+import configureStore, { history } from './configureStore'
 
 const initialState = Immutable.Map()
-const composeEnhancer = window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__ || compose
-const store = createStore(
-  rootReducer(history),
-  initialState,
-  composeEnhancer(
-    applyMiddleware(
-      routerMiddleware(history),
-    ),
-  ),
-)
-
+const store = configureStore(initialState)
 const render = () => {
   ReactDOM.render(
     <AppContainer>
@@ -41,10 +26,5 @@ if (module.hot) {
   // Reload components
   module.hot.accept('./App', () => {
     render()
-  })
-
-  // Reload reducers
-  module.hot.accept('./reducers', () => {
-    store.replaceReducer(rootReducer(history))
   })
 }
