+++ examples/typescript/src/index.tsx
@@ -1,24 +1,9 @@
 import { AppContainer } from 'react-hot-loader'
-import { applyMiddleware, compose, createStore } from 'redux'
-import { createBrowserHistory } from 'history'
-import { routerMiddleware } from 'connected-react-router'
 import { Provider } from 'react-redux'
 import React from 'react'
 import ReactDOM from 'react-dom'
 import App from './App'
-import rootReducer from './reducers'
-
-const history = createBrowserHistory()
-
-const composeEnhancer: typeof compose = (window as any).__REDUX_DEVTOOLS_EXTENSION_COMPOSE__ || compose
-const store = createStore(
-  rootReducer(history),
-  composeEnhancer(
-    applyMiddleware(
-      routerMiddleware(history),
-    ),
-  ),
-)
+import configureStore, { history } from './configureStore'
 
 const render = () => {
   ReactDOM.render(
@@ -38,10 +23,5 @@ if (module.hot) {
   // Reload components
   module.hot.accept('./App', () => {
     render()
-  })
-
-  // Reload reducers
-  module.hot.accept('./reducers', () => {
-    store.replaceReducer(rootReducer(history))
   })
 }
